pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "spft"
version = "1.0.0"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]
description = "SparseLoRA"
authors = [
    {name = "Samir Khaki",email = "samir.khaki@mail.utoronto.ca"}
]

[project.optional-dependencies]
base = [
    "setuptools (==75.1.0)",
    "wheel (==0.44.0)",
    "pip (==24.2)",
    "urllib3 (==2.2.3)",
    "typing-inspection (==0.4.0)",
    "typing-extensions (==4.12.2)",
    "smmap (==5.0.1)",
    "six (==1.16.0)",
    "setproctitle (==1.3.4)",
    "pyyaml (==6.0.2)",
    "psutil (==6.1.0)",
    "protobuf (==5.28.3)",
    "platformdirs (==4.3.6)",
    "idna (==3.10)",
    "click (==8.1.7)",
    "charset-normalizer (==3.4.0)",
    "certifi (==2024.8.30)",
    "sentry-sdk (==2.18.0)",
    "requests (==2.32.3)",
    "gitdb (==4.0.11)",
    "docker-pycreds (==0.4.0)",
    "gitpython (==3.1.43)",
    "wandb (==0.18.7)",
    "tqdm (==4.67.0)",
    "safetensors (==0.4.5)",
    "regex (==2024.11.6)",
    "packaging (==24.2)",
    "numpy (==2.1.3)",
    "frozenlist (==1.5.0)",
    "filelock (==3.16.1)",
    "huggingface-hub (==0.26.2)",
    "tokenizers (==0.20.3)",
    "transformers (==4.46.3)",
    "mpmath (==1.3.0)",
    "triton (==3.1.0)",
    "sympy (==1.13.1)",
    "nvidia-nvtx-cu12 (==12.4.127)",
    "nvidia-nvjitlink-cu12 (==12.4.127)",
    "nvidia-nccl-cu12 (==2.21.5)",
    "nvidia-curand-cu12 (==10.3.5.147)",
    "nvidia-cufft-cu12 (==11.2.1.3)",
    "nvidia-cuda-runtime-cu12 (==12.4.127)",
    "nvidia-cuda-nvrtc-cu12 (==12.4.127)",
    "nvidia-cuda-cupti-cu12 (==12.4.127)",
    "nvidia-cublas-cu12 (==12.4.5.8)",
    "networkx (==3.4.2)",
    "markupsafe (==3.0.2)",
    "nvidia-cusparse-cu12 (==12.3.1.170)",
    "nvidia-cudnn-cu12 (==9.1.0.70)",
    "jinja2 (==3.1.4)",
    "nvidia-cusolver-cu12 (==11.6.1.9)",
    "torch (==2.5.1)",
    "accelerate (==1.1.1)",
    "peft (==0.13.2)",
    "pytz (==2024.2)",
    "xxhash (==3.5.0)",
    "tzdata (==2024.2)",
    "python-dateutil (==2.9.0.post0)",
    "pyarrow (==18.0.0)",
    "propcache (==0.2.0)",
    "multidict (==6.1.0)",
    "fsspec (==2024.9.0)",
    "dill (==0.3.8)",
    "attrs (==24.2.0)",
    "async-timeout (==5.0.1)",
    "aiohappyeyeballs (==2.4.3)",
    "yarl (==1.17.2)",
    "pandas (==2.2.3)",
    "multiprocess (==0.70.16)",
    "aiosignal (==1.3.1)",
    "aiohttp (==3.11.6)",
    "datasets (==3.1.0)",
    "tabulate (==0.9.0)",
    "einops (==0.8.0)",
    "sentencepiece (==0.2.0)",
    "et-xmlfile (==2.0.0)",
    "openpyxl (==3.1.5)",
    "bitsandbytes (==0.45.5)",
    "annotated-types (==0.7.0)",
    "liger-kernel (==0.5.9)",
    "pydantic (==2.11.4)",
    "pydantic-core (==2.33.2)",
    "userpath (==1.9.2)",
    "numpy (==2.1.3)",
    "trl (==0.17.0)",
]

unsloth = [
    "setuptools (==75.1.0)",
    "wheel (==0.44.0)",
    "pip (==24.2)",
    "urllib3 (==2.2.3)",
    "typing-inspection (==0.4.0)",
    "typing-extensions (==4.12.2)",
    "smmap (==5.0.1)",
    "six (==1.16.0)",
    "setproctitle (==1.3.4)",
    "pyyaml (==6.0.2)",
    "psutil (==6.1.0)",
    "protobuf (==5.28.3)",
    "platformdirs (==4.3.6)",
    "idna (==3.10)",
    "click (==8.1.7)",
    "charset-normalizer (==3.4.0)",
    "certifi (==2024.8.30)",
    "sentry-sdk (==2.18.0)",
    "requests (==2.32.3)",
    "gitdb (==4.0.11)",
    "docker-pycreds (==0.4.0)",
    "gitpython (==3.1.43)",
    "wandb (==0.18.7)",
    "tqdm (==4.67.0)",
    "safetensors (==0.4.5)",
    "regex (==2024.11.6)",
    "packaging (==24.2)",
    "numpy (==2.1.3)",
    "frozenlist (==1.5.0)",
    "filelock (==3.16.1)",
    "huggingface-hub (==0.33.0)",
    "tokenizers (==0.21.1)",
    "transformers (==4.51.3)",
    "mpmath (==1.3.0)",
    "triton (==3.1.0)",
    "sympy (==1.13.1)",
    "nvidia-nvtx-cu12 (==12.4.127)",
    "nvidia-nvjitlink-cu12 (==12.4.127)",
    "nvidia-nccl-cu12 (==2.21.5)",
    "nvidia-curand-cu12 (==10.3.5.147)",
    "nvidia-cufft-cu12 (==11.2.1.3)",
    "nvidia-cuda-runtime-cu12 (==12.4.127)",
    "nvidia-cuda-nvrtc-cu12 (==12.4.127)",
    "nvidia-cuda-cupti-cu12 (==12.4.127)",
    "nvidia-cublas-cu12 (==12.4.5.8)",
    "networkx (==3.4.2)",
    "markupsafe (==3.0.2)",
    "nvidia-cusparse-cu12 (==12.3.1.170)",
    "nvidia-cudnn-cu12 (==9.1.0.70)",
    "jinja2 (==3.1.4)",
    "nvidia-cusolver-cu12 (==11.6.1.9)",
    "torch (==2.5.1)",
    "accelerate (==1.1.1)",
    "peft (==0.13.2)",
    "pytz (==2024.2)",
    "xxhash (==3.5.0)",
    "tzdata (==2024.2)",
    "python-dateutil (==2.9.0.post0)",
    "pyarrow (==18.0.0)",
    "propcache (==0.2.0)",
    "multidict (==6.1.0)",
    "fsspec (==2024.9.0)",
    "dill (==0.3.8)",
    "attrs (==24.2.0)",
    "async-timeout (==5.0.1)",
    "aiohappyeyeballs (==2.4.3)",
    "yarl (==1.17.2)",
    "pandas (==2.2.3)",
    "multiprocess (==0.70.16)",
    "aiosignal (==1.3.1)",
    "aiohttp (==3.11.6)",
    "datasets (==3.1.0)",
    "tabulate (==0.9.0)",
    "einops (==0.8.0)",
    "sentencepiece (==0.2.0)",
    "et-xmlfile (==2.0.0)",
    "openpyxl (==3.1.5)",
    "bitsandbytes (==0.45.5)",
    "annotated-types (==0.7.0)",
    "liger-kernel (==0.5.9)",
    "pydantic (==2.11.4)",
    "pydantic-core (==2.33.2)",
    "userpath (==1.9.2)",
    "numpy (==2.1.3)",
    "trl (==0.17.0)",
]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120

[tool.setuptools]
packages = ["spft", "tools"]